merge code

+++ app/controllers/users_controller.ts
@@ -1,21 +1,33 @@
 import type { HttpContext } from '@adonisjs/core/http'
-import app from '@adonisjs/core/services/app'
-import { cuid } from '@adonisjs/core/helpers'
 import vine from '@vinejs/vine'
 import BaseController from '#core/base_controller'
 import User from '#models/user'
 import { uniqueRule } from '#rules/unique'
 import { existsRule } from '#rules/exists'
-import fs from 'node:fs'
 
 export default class UsersController extends BaseController {
   /**
    * Display a list of resource
    */
-  async index() {
-    const data = await User.all()
+  async index({ request }: HttpContext) {
 
-    this.response('Users retrieved successfully', data)
+    const params = request.all()
+
+    let query: any = {
+      table_and_join: 'from users',
+      field_show: [
+        'id',
+        'name',
+        'email',
+        'user_role_id',
+        'user_status_id'
+      ],
+      field_search: ['email', 'name'],
+      pagination: true,
+    }
+
+    const result = await this.query.generate(params, query, this.db)
+    this.responseList('Users retrieved successfully', result)
   }
 
   /**
@@ -25,14 +37,8 @@ export default class UsersController extends BaseController {
     const payload = request.body()
     const validator = vine.compile(
       vine.object({
-        userStatusId: vine.number().use(existsRule({ table: 'user_statuses', column: 'id' })),
-        userRoleId: vine.number().use(existsRule({ table: 'user_roles', column: 'id' })),
-        username: vine.string().use(
-          uniqueRule({
-            table: 'users',
-            column: 'username',
-          })
-        ),
+        user_status_id: vine.number().use(existsRule({ table: 'user_statuses', column: 'id' })),
+        user_role_id: vine.number().use(existsRule({ table: 'user_roles', column: 'id' })),
         password: vine
           .string()
           .minLength(8)
@@ -42,7 +48,6 @@ export default class UsersController extends BaseController {
           })
           .optional(),
         name: vine.string(),
-        photo: vine.string().optional(),
         email: vine
           .string()
           .email()
@@ -51,27 +56,11 @@ export default class UsersController extends BaseController {
     )
     const output = await validator.validate(payload)
 
-    const photo = request.file('photo', {
-      size: '2mb',
-      extnames: ['jpg', 'png', 'jpeg'],
-    })
 
-    // upload photo
-    if (photo) {
-      if (!photo.isValid) {
-        return this.responseError('Validation error', 422, photo.errors)
-      }
-
-      await photo.move(app.makePath('uploads/user-photo'), {
-        name: `${cuid()}.${photo.extname}`,
-      })
-
-      output.photo = photo.fileName!
-    }
 
     const data = await User.create(output)
 
-    this.response('User created successfully', data)
+    this.response('User created successfully', { item: data })
   }
 
   /**
@@ -83,7 +72,7 @@ export default class UsersController extends BaseController {
     await data.load('user_status')
     await data.load('posts')
 
-    this.response('User retrieved successfully', data)
+    this.response('User retrieved successfully', { item: data })
   }
 
   /**
@@ -94,16 +83,8 @@ export default class UsersController extends BaseController {
     const validator = vine.compile(
       vine.object({
         id: vine.number().use(existsRule({ table: 'users', column: 'id' })),
-        userStatusId: vine.number().use(existsRule({ table: 'user_statuses', column: 'id' })),
-        userRoleId: vine.number().use(existsRule({ table: 'user_roles', column: 'id' })),
-        username: vine.string().use(
-          uniqueRule({
-            table: 'users',
-            column: 'username',
-            except: params.id,
-            exceptColumn: 'id',
-          })
-        ),
+        user_status_id: vine.number().use(existsRule({ table: 'user_statuses', column: 'id' })),
+        user_role_id: vine.number().use(existsRule({ table: 'user_roles', column: 'id' })),
         password: vine
           .string()
           .minLength(8)
@@ -113,7 +94,6 @@ export default class UsersController extends BaseController {
           })
           .optional(),
         name: vine.string(),
-        photo: vine.string().optional(),
         email: vine
           .string()
           .email()
@@ -124,35 +104,9 @@ export default class UsersController extends BaseController {
     )
     const output = await validator.validate({ id: params.id, ...payload })
     const data = await User.findOrFail(params.id)
-
-    const photo = request.file('photo', {
-      size: '2mb',
-      extnames: ['jpg', 'png', 'jpeg'],
-    })
-
-    // upload photo
-    if (photo) {
-      if (!photo.isValid) {
-        return this.responseError('Validation error', 422, photo.errors)
-      }
-
-      // delete old file
-      if (data.photo) {
-        fs.unlink(app.makePath(`uploads/user-photo/${data.photo}`), (err) => {
-          if (err) console.error('Error removing file:', err)
-        })
-      }
-
-      await photo.move(app.makePath('uploads/user-photo'), {
-        name: `${cuid()}.${photo.extname}`,
-      })
-
-      output.photo = photo.fileName!
-    }
-
     await data?.merge(output).save()
 
-    this.response('User updated successfully', data)
+    this.response('User updated successfully', { item: data })
   }
 
   /**
@@ -160,6 +114,9 @@ export default class UsersController extends BaseController {
    */
   async destroy({ params }: HttpContext) {
     const data = await User.findOrFail(params.id)
+    if (data.id === 1) {
+      return this.responseError('Cannot delete super admin', 412)
+    }
     await data?.delete()
 
     this.response('User deleted successfully')
